Restructured the old layout
