+++ pilot/HistoryTableViewCell.swift
@@ -1,0 +1,24 @@
+//
+//  HistoryTableViewCell.swift
+//  pilot
+//
+//  Created by Nick Eckert on 11/19/17.
+//  Copyright © 2017 sanction. All rights reserved.
+//
+
+import UIKit
+
+class HistoryTableViewCell: UITableViewCell {
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+
+}
