+++ pilot/HistoryViewController.swift
@@ -1,0 +1,29 @@
+//
+//  HistoryViewController.swift
+//  pilot
+//
+//  Created by Nick Eckert on 11/19/17.
+//  Copyright © 2017 sanction. All rights reserved.
+//
+
+import UIKit
+
+class HistoryViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        styleUI()
+        loadUI()
+    }
+
+  func styleUI() {
+
+  }
+
+  func loadUI() {
+
+  }
+
+  
+}
