+++ pilot/HomeTabBarController.swift
@@ -1,0 +1,26 @@
+//
+//  HomeTabBarController.swift
+//  pilot
+//
+//  Created by Nick Eckert on 11/19/17.
+//  Copyright © 2017 sanction. All rights reserved.
+//
+
+import UIKit
+
+class HomeTabBarController: UITabBarController {
+
+  override func viewDidLoad() {
+    self.delegate = self
+  }
+
+  
+}
+
+extension HomeTabBarController: UITabBarControllerDelegate {
+
+  func tabBarController(_ tabBarController: UITabBarController, didSelect viewController: UIViewController) {
+    // TabBar item was selected with a viewController
+  }
+
+}
